+++ src/anemoi/datasets/dates/__init__.py
@@ -12,6 +12,7 @@ import warnings
 # from anemoi.utils.dates import as_datetime
 from anemoi.utils.dates import DateTimes
 from anemoi.utils.dates import as_datetime
+from anemoi.utils.dates import frequency_to_string
 from anemoi.utils.dates import frequency_to_timedelta
 from anemoi.utils.hindcasts import HindcastDatesTimes
 from anemoi.utils.humanize import print_dates
